adds grid cube to Canvas and lowers grid and takeoffLine by gridEdgeLength/2. Also modularizes skyboxes and drone3DModel. closes #79
