+++ src/assets/skybox/nebula/back.png
