+++ src/assets/skybox/nebula/front.png
