+++ src/assets/skybox/nebula/right.png
